LayoutTests/fast/js/xhtml-serialize-expected.txt

This test checks whether serialized invalid XHTML is valid XML (for bug 9901).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS doc.firstChild.firstChild.nodeValue is "123"
PASS str is '<input xmlns="http://www.w3.org/1999/xhtml">123</input>'
PASS successfullyParsed is true

TEST COMPLETE
